Rename `lib/` to `src/` and `http/` to `public/`
The new directory names are more commonly used among PHP projects. It's
also the default names when creating a new project with Symfony.
