+++ config/config.default.php
@@ -22,7 +22,7 @@ $CONFIG = array(
         "version" => "3.0",
         "title" => "The Mesa drivers matrix",
         "description" => "Show Mesa progress for the OpenGL, OpenGL ES, Vulkan and OpenCL drivers implementations into an easy to read HTML page.",
-        "xml_file" => "http/features.xml",
+        "xml_file" => "public/features.xml",
         "project_url" => "https://github.com/MightyCreak/mesamatrix",
         "commitlog_length" => 10,
         "authors" => array(
