+++ mesamatrixctl
@@ -19,7 +19,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-require_once "lib/base.php";
+require_once "src/base.php";
 
 ini_set('display_errors', '1');
