+++ public/rss.php
@@ -3,7 +3,8 @@
 /*
  * This file is part of mesamatrix.
  *
- * Copyright (C) 2014-2017 Romain "Creak" Failliot.
+ * Copyright (C) 2020 Romain "Creak" Failliot.
+ * Copyright (C) 2015 Robin McCorkell <rmccorkell@karoshi.org.uk>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -19,7 +20,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-require_once "../lib/base.php";
+require_once "../src/base.php";
 
-$controller = new Mesamatrix\Controller\DriversController();
-$controller->writeHtml();
+$rssGenerator = new Mesamatrix\Controller\RssController();
+$rssGenerator->run();
