+++ src/Git/Commit.php
