+++ src/Parser/Hints.php
